+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/UTxO.hs
@@ -32,6 +32,7 @@ module Cardano.Ledger.Shelley.UTxO (
 import Cardano.Ledger.Address (accountAddressCredentialL, bootstrapKeyHash)
 import Cardano.Ledger.BaseTypes (StrictMaybe (..))
 import Cardano.Ledger.Coin (Coin (..))
+import Cardano.Ledger.Compactible (fromCompact)
 import Cardano.Ledger.Credential (Credential (..), credKeyHashWitness, credScriptHash)
 import Cardano.Ledger.Keys (
   GenDelegs (..),
@@ -42,13 +43,7 @@ import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.Era (ShelleyEra)
 import Cardano.Ledger.Shelley.PParams (ProposedPPUpdates (ProposedPPUpdates), Update (..))
 import Cardano.Ledger.Shelley.State ()
-import Cardano.Ledger.State (
-  EraCertState (..),
-  StakePoolParams (..),
-  dsGenDelegs,
-  lookupDepositDState,
-  psStakePoolsL,
- )
+import Cardano.Ledger.State
 import Cardano.Ledger.State as UTxO (
   CanGetUTxO (..),
   CanSetUTxO (..),
@@ -116,15 +111,13 @@ getShelleyScriptsNeeded u txBody =
 shelleyConsumed ::
   (EraUTxO era, EraCertState era) =>
   PParams era ->
-  CertState era ->
+  Accounts era ->
   UTxO era ->
   TxBody l era ->
   Value era
-shelleyConsumed pp certState =
-  getConsumedValue
-    pp
-    (lookupDepositDState $ certState ^. certDStateL)
-    (const Nothing)
+shelleyConsumed pp accounts =
+  getConsumedValue pp $
+    fmap (fromCompact . (^. depositAccountStateL)) . (`lookupAccountState` accounts)
 
 -- | Compute the lovelace which are created by the transaction
 -- For eras before Conway, VState is expected to have an empty Map for vsDReps, and so deposit summed up is zero.
@@ -164,7 +157,7 @@ getConsumedCoin pp lookupRefund utxo txBody =
     <> refunds
     <> withdrawals
   where
-    refunds = getTotalRefundsTxBody pp lookupRefund (const Nothing) txBody
+    refunds = getTotalRefundsTxBody pp lookupRefund txBody
     withdrawals = fold . unWithdrawals $ txBody ^. withdrawalsTxBodyL
 
 newtype ShelleyScriptsNeeded era = ShelleyScriptsNeeded (Set ScriptHash)
@@ -175,7 +168,7 @@ instance EraUTxO ShelleyEra where
 
   consumed = shelleyConsumed
 
-  getConsumedValue pp lookupKeyDeposit _ = getConsumedCoin pp lookupKeyDeposit
+  getConsumedValue = getConsumedCoin
 
   getProducedValue = shelleyProducedValue